+++ src/train-container/train-book/index.less
@@ -8,6 +8,9 @@
     width: 256px;
     flex-grow: 0;
     flex-shrink: 0;
+    a {
+      color: white;
+    }
   }
   .stage-container {
     height: 100%;
