+++ src/train-container/train-info/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default class TrainInfo extends React.Component {
+  render(){
+    return (
+      <div>我的培训室</div>
+    )
+  }
+}
