+++ src/train-container/train-sheet/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default class TrainSheet extends React.Component {
+  render() {
+    return <div>我的统计报表</div>;
+  }
+}
